separar styled-components
