+++ src/pages/Home/Home.js
@@ -1,0 +1,172 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+    
+`
+
+export const Header = styled.div`
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height: 60px;
+    background-color: #1960B1;
+
+  > * {
+    &:first-child{
+        img{
+            height: 40px;
+            margin-left: 16px;
+            margin-right: 30px;
+        }
+    }
+  }
+
+  > * {
+    &:nth-child(2){
+        input{
+            height: 40px;
+            width: 230px;
+            border: none;
+            border-radius: 6px;
+        }
+    }
+  }
+
+  > * {
+    &:nth-child(3){
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        width: 70px;
+    }
+  }
+`
+
+export const Slides = styled.div`
+
+`
+
+export const Menu = styled.div`
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    width: 100%;
+    height: calc(100vh - 60px);
+    background-color: #1960B1;
+    padding: 20px 16px 20px 16px;
+
+    > * {
+    &:first-child{
+        > * {
+        &:first-child{
+            display: flex;
+            align-items: center;
+            color: #ffffff;
+
+            > * {
+                &:first-child{
+                    display: flex;
+                    justify-content: center;
+                    align-items: center;
+                    width: 30px;
+                    height: 30px;
+                    border-radius: 50%;
+                    border: 1px solid #ffffff;
+                    margin-right: 8px;
+                }
+            }
+        }
+    }
+
+    > * {
+        &:nth-child(2){
+            display: flex;
+            justify-content: space-between;
+            width: 100px;
+            color: #ffffff;
+            margin-top: 16px;
+            margin-left: 12px;
+        }
+    }
+
+    > * {
+        &:nth-child(3){
+            display: flex;
+            justify-content: space-between;
+            width: 56px;
+            color: #ffffff;
+            margin-top: 16px;
+            margin-left: 12px;
+        }
+    }
+    }
+  }
+
+  > * {
+    &:last-child{
+        display: flex;
+        flex-direction: column;
+
+        button{
+            height: 40px;
+            background-color: #EB881E;
+            border: none;
+            border-radius: 6px;
+            margin-bottom: 6px;
+            color: #ffffff;
+            font-size: 20px;
+        }
+    }
+  }
+`
+
+export const Section = styled.div`
+    > * {
+        &:nth-child(2){
+            display: flex;
+            justify-content: space-between;
+            align-items: center;
+            height: 40px;
+            width: 100%;
+            background-color: #E65F0D;
+            color:#ffffff;
+            padding-left: 20px;
+            padding-right: 20px;
+
+            > * {
+                &:nth-child(2){
+                    display: flex;
+                    justify-content: space-between;
+                    align-items: center;
+                    width: 110px;
+                }
+            }
+        }
+    }
+`
+
+export const Contents = styled.div`
+    display: flex;
+    flex-direction: column;
+    background-color: #FAFAFB;
+    width: 100%;
+    height: calc(100vh - 260px);
+    padding: 20px 16px 20px 16px;
+`
+
+export const Products = styled.div`
+    display: flex;
+    overflow-x: scroll;
+    scroll-behavior: auto;
+`
+
+export const Title = styled.div`
+    display: flex;
+    font-size: 20px;
+    color: #000000;
+    font-weight: bolder;
+    margin-top: 20px;
+    margin-bottom: 20px;
+`
+
